deploy/backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.10.0
    # via
    #   django
    #   django-cors-headers
colorama==0.4.6
    # via pytest
django==5.0.2
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   djangorestframework
django-cors-headers==4.9.0
    # via -r requirements.in
django-environ==0.12.0
    # via -r requirements.in
django-ratelimit==4.1.0
    # via -r requirements.in
djangorestframework==3.16.1
    # via -r requirements.in
factory-boy==3.3.3
    # via
    #   -r requirements.in
    #   pytest-factoryboy
faker==40.15.0
    # via factory-boy
inflection==0.5.1
    # via pytest-factoryboy
iniconfig==2.3.0
    # via pytest
packaging==26.1
    # via
    #   pytest
    #   pytest-factoryboy
pillow==11.3.0
    # via -r requirements.in
pluggy==1.6.0
    # via pytest
psycopg[binary]==3.3.3
    # via -r requirements.in
psycopg-binary==3.3.3
    # via psycopg
pygments==2.20.0
    # via pytest
pytest==8.4.2
    # via
    #   -r requirements.in
    #   pytest-django
    #   pytest-factoryboy
pytest-django==4.12.0
    # via -r requirements.in
pytest-factoryboy==2.8.1
    # via -r requirements.in
sqlparse==0.5.3
    # via django
typing-extensions==4.15.0
    # via
    #   psycopg
    #   pytest-factoryboy
tzdata==2026.1
    # via
    #   django
    #   faker
    #   psycopg
